fixing when coverage is emitted to handle larger storybook instances
