+++ packages/storycap/src/node/screenshot-service.ts
@@ -3,8 +3,6 @@ import { CapturingBrowser } from './capturing-browser';
 import { FileSystem } from './file';
 import { Logger } from './logger';
 import { VariantKey } from '../shared/types';
-import { JSCoverageEntry } from 'puppeteer-core';
-import mergeCoverages from './coverage';
 
 function createRequest({
   story,
@@ -50,7 +48,6 @@ export type ScreenshotServiceOptions = {
   workers: CapturingBrowser[];
   fileSystem: FileSystem;
   stories: Story[];
-  coverage: string;
 };
 
 /**
@@ -66,9 +63,7 @@ export function createScreenshotService({
   logger,
   stories,
   workers,
-  coverage: coveragePath,
 }: ScreenshotServiceOptions): ScreenshotService {
-  const coverages: JSCoverageEntry[] = [];
   const service = createExecutionService(
     workers,
     stories.map(story => createRequest({ story })),
@@ -77,7 +72,7 @@ export function createScreenshotService({
         // Delegate the request to the worker.
         const [result, elapsedTime] = await time(worker.screenshot(rid, story, variantKey, count));
 
-        const { succeeded, buffer, coverage, variantKeysToPush, defaultVariantSuffix } = result;
+        const { succeeded, buffer, variantKeysToPush, defaultVariantSuffix } = result;
 
         // Queue retry request if the request was not succeeded.
         // Worker throws `ScreenshotTimeoutError` if the queued request continues failed and the count exceeds the threhold.
@@ -85,10 +80,6 @@ export function createScreenshotService({
 
         // Queue screenshot requests for additional variants.
         variantKeysToPush.forEach(variantKey => push(createRequest({ story, variantKey })));
-
-        if (coverage) {
-          coverages.push(...coverage);
-        }
 
         if (buffer) {
           const suffix = variantKey.isDefault && defaultVariantSuffix ? [defaultVariantSuffix] : variantKey.keys;
@@ -99,18 +90,6 @@ export function createScreenshotService({
       },
   );
   return {
-    execute: () =>
-      service
-        .execute()
-        .then(async captured => {
-          if (coveragePath && coverages.length) {
-            logger.log(`coverage data stored: ${logger.color.magenta('coverage.json')}.`);
-            const merged = await mergeCoverages(coverages, coveragePath);
-            const buffer = Buffer.from(JSON.stringify(merged));
-            await fileSystem.saveFile('coverage.json', buffer);
-          }
-          return captured;
-        })
-        .then(captured => captured.filter(c => !!c).length),
+    execute: () => service.execute().then(captured => captured.filter(c => !!c).length),
   };
 }
